Boltzmann sampling function added in utils/sampling to remove duplicate code, reshuffling of other sampling methods (that don't take an acqf)
