+++ botorch/utils/sampling.py
@@ -21,18 +21,27 @@ import warnings
 from abc import ABC, abstractmethod
 from collections.abc import Callable, Generator, Iterable
 from contextlib import contextmanager
+from math import ceil
 from typing import TYPE_CHECKING
 
 import numpy as np
 import numpy.typing as npt
 import scipy
+
 import torch
-from botorch.exceptions.errors import BotorchError, InfeasibilityError
+
+from botorch.exceptions.errors import (
+    BotorchError,
+    BotorchTensorDimensionError,
+    InfeasibilityError,
+)
 from botorch.exceptions.warnings import UserInputWarning
 from botorch.sampling.qmc import NormalQMCEngine
-from botorch.utils.transforms import unnormalize
+
+from botorch.utils.transforms import normalize, unnormalize
 from scipy.spatial import Delaunay, HalfspaceIntersection
 from torch import LongTensor, Tensor
+from torch.distributions import Normal
 from torch.quasirandom import SobolEngine
 
 
@@ -1061,3 +1070,166 @@ def optimize_posterior_samples(
 
     f_opt = paths(X_opt.unsqueeze(-2)).squeeze(-2)
     return X_opt, f_opt
+
+
+def boltzmann_sample(
+    function_values: Tensor,
+    num_samples: int,
+    eta: float,
+    replacement: bool = False,
+    temp_decrease: float = 0.5,
+):
+    """
+    Perform Boltzmann sampling from a set of function values, weighted by the
+    exponentiated difference between function values and their standardized mean.
+
+    Args:
+        function_values: A [batch_shape] x N  tensor of function values.
+        num_samples: The number of samples (restarts) to draw.
+        eta: The Boltzmann temperature, controls the sharpness of the weighting. If the
+            temperature is too high, causing NaN values, the eta parameter is
+            succesively decreased by 'temp_decrease'.
+        replacement: If True, samples are drawn with replacement, allowing duplicates.
+        temp_decrease: The rate at which temperature decreases in case of inf weights.
+    Returns:
+        A [batch_shape] x num_samples tensor of indices of sampled positions.
+    """
+    # standardize() is not used since function_values may be more than 2D
+    norm_weights = (
+        function_values - function_values.mean(dim=-1, keepdim=True)
+    ) / function_values.std(dim=-1, keepdim=True)
+    weights = torch.exp(eta * norm_weights)
+    while torch.isinf(weights).any():
+        eta *= temp_decrease
+        weights = torch.exp(eta * norm_weights)
+
+    return batched_multinomial(
+        weights=weights, num_samples=num_samples, replacement=replacement
+    )
+
+
+def sample_truncated_normal_perturbations(
+    X: Tensor,
+    n_discrete_points: int,
+    sigma: float,
+    bounds: Tensor,
+    qmc: bool = True,
+) -> Tensor:
+    r"""Sample points around `X`.
+
+    Sample perturbed points around `X` such that the added perturbations
+    are sampled from N(0, sigma^2 I) and truncated to be within [0,1]^d.
+
+    Args:
+        X: A `n x d`-dim tensor starting points.
+        n_discrete_points: The number of points to sample.
+        sigma: The standard deviation of the additive gaussian noise for
+            perturbing the points.
+        bounds: A `2 x d`-dim tensor containing the bounds.
+        qmc: A boolean indicating whether to use qmc.
+
+    Returns:
+        A `n_discrete_points x d`-dim tensor containing the sampled points.
+    """
+    X = normalize(X, bounds=bounds)
+    d = X.shape[1]
+    # sample points from N(X_center, sigma^2 I), truncated to be within
+    # [0, 1]^d.
+    if X.shape[0] > 1:
+        rand_indices = torch.randint(X.shape[0], (n_discrete_points,), device=X.device)
+        X = X[rand_indices]
+    if qmc:
+        std_bounds = torch.zeros(2, d, dtype=X.dtype, device=X.device)
+        std_bounds[1] = 1
+        u = draw_sobol_samples(bounds=std_bounds, n=n_discrete_points, q=1).squeeze(1)
+    else:
+        u = torch.rand((n_discrete_points, d), dtype=X.dtype, device=X.device)
+    # compute bounds to sample from
+    a = -X
+    b = 1 - X
+    # compute z-score of bounds
+    alpha = a / sigma
+    beta = b / sigma
+    normal = Normal(0, 1)
+    cdf_alpha = normal.cdf(alpha)
+    # use inverse transform
+    perturbation = normal.icdf(cdf_alpha + u * (normal.cdf(beta) - cdf_alpha)) * sigma
+    # add perturbation and clip points that are still outside
+    perturbed_X = (X + perturbation).clamp(0.0, 1.0)
+    return unnormalize(perturbed_X, bounds=bounds)
+
+
+def sample_perturbed_subset_dims(
+    X: Tensor,
+    bounds: Tensor,
+    n_discrete_points: int,
+    sigma: float = 1e-1,
+    qmc: bool = True,
+    prob_perturb: float | None = None,
+) -> Tensor:
+    r"""Sample around `X` by perturbing a subset of the dimensions.
+
+    By default, dimensions are perturbed with probability equal to
+    `min(20 / d, 1)`. As shown in [Regis]_, perturbing a small number
+    of dimensions can be beneificial. The perturbations are sampled
+    from N(0, sigma^2 I) and truncated to be within [0,1]^d.
+
+    Args:
+        X: A `n x d`-dim tensor starting points. `X`
+            must be normalized to be within `[0, 1]^d`.
+        bounds: The bounds to sample perturbed values from
+        n_discrete_points: The number of points to sample.
+        sigma: The standard deviation of the additive gaussian noise for
+            perturbing the points.
+        qmc: A boolean indicating whether to use qmc.
+        prob_perturb: The probability of perturbing each dimension. If omitted,
+            defaults to `min(20 / d, 1)`.
+
+    Returns:
+        A `n_discrete_points x d`-dim tensor containing the sampled points.
+
+    """
+    if bounds.ndim != 2:
+        raise BotorchTensorDimensionError("bounds must be a `2 x d`-dim tensor.")
+    elif X.ndim != 2:
+        raise BotorchTensorDimensionError("X must be a `n x d`-dim tensor.")
+    d = bounds.shape[-1]
+    if prob_perturb is None:
+        # Only perturb a subset of the features
+        prob_perturb = min(20.0 / d, 1.0)
+
+    if X.shape[0] == 1:
+        X_cand = X.repeat(n_discrete_points, 1)
+    else:
+        rand_indices = torch.randint(X.shape[0], (n_discrete_points,), device=X.device)
+        X_cand = X[rand_indices]
+    pert = sample_truncated_normal_perturbations(
+        X=X_cand,
+        n_discrete_points=n_discrete_points,
+        sigma=sigma,
+        bounds=bounds,
+        qmc=qmc,
+    )
+
+    # find cases where we are not perturbing any dimensions
+    mask = (
+        torch.rand(
+            n_discrete_points,
+            d,
+            dtype=bounds.dtype,
+            device=bounds.device,
+        )
+        <= prob_perturb
+    )
+    ind = (~mask).all(dim=-1).nonzero()
+    # perturb `n_perturb` of the dimensions
+    n_perturb = ceil(d * prob_perturb)
+    perturb_mask = torch.zeros(d, dtype=mask.dtype, device=mask.device)
+    perturb_mask[:n_perturb].fill_(1)
+    # TODO: use batched `torch.randperm` when available:
+    # https://github.com/pytorch/pytorch/issues/42502
+    for idx in ind:
+        mask[idx] = perturb_mask[torch.randperm(d, device=bounds.device)]
+    # Create candidate points
+    X_cand[mask] = pert[mask]
+    return X_cand
