+++ test/optim/test_initializers.py
@@ -1266,111 +1266,6 @@ class TestGenValueFunctionInitialConditions(BotorchTestCase):
 
 
 class TestSampleAroundBest(BotorchTestCase):
-    def test_sample_truncated_normal_perturbations(self):
-        tkwargs = {"device": self.device}
-        n_discrete_points = 5
-        _bounds = torch.ones(2, 4)
-        _bounds[1] = 2
-        for dtype in (torch.float, torch.double):
-            tkwargs["dtype"] = dtype
-            bounds = _bounds.to(**tkwargs)
-            for n_best in (1, 2):
-                X = 1 + torch.rand(n_best, 4, **tkwargs)
-                # basic test
-                perturbed_X = sample_truncated_normal_perturbations(
-                    X=X,
-                    n_discrete_points=n_discrete_points,
-                    sigma=4,
-                    bounds=bounds,
-                    qmc=False,
-                )
-                self.assertEqual(perturbed_X.shape, torch.Size([n_discrete_points, 4]))
-                self.assertTrue((perturbed_X >= 1).all())
-                self.assertTrue((perturbed_X <= 2).all())
-                # test qmc
-                with mock.patch(
-                    "botorch.optim.initializers.draw_sobol_samples",
-                    wraps=draw_sobol_samples,
-                ) as mock_sobol:
-                    perturbed_X = sample_truncated_normal_perturbations(
-                        X=X,
-                        n_discrete_points=n_discrete_points,
-                        sigma=4,
-                        bounds=bounds,
-                        qmc=True,
-                    )
-                    mock_sobol.assert_called_once()
-                self.assertEqual(perturbed_X.shape, torch.Size([n_discrete_points, 4]))
-                self.assertTrue((perturbed_X >= 1).all())
-                self.assertTrue((perturbed_X <= 2).all())
-
-    def test_sample_perturbed_subset_dims(self):
-        tkwargs = {"device": self.device}
-        n_discrete_points = 5
-
-        # test that errors are raised
-        with self.assertRaises(BotorchTensorDimensionError):
-            sample_perturbed_subset_dims(
-                X=torch.zeros(1, 1),
-                n_discrete_points=1,
-                sigma=1e-3,
-                bounds=torch.zeros(1, 2, 1),
-            )
-        with self.assertRaises(BotorchTensorDimensionError):
-            sample_perturbed_subset_dims(
-                X=torch.zeros(1, 1, 1),
-                n_discrete_points=1,
-                sigma=1e-3,
-                bounds=torch.zeros(2, 1),
-            )
-        for dtype in (torch.float, torch.double):
-            for n_best in (1, 2):
-                tkwargs["dtype"] = dtype
-                bounds = torch.zeros(2, 21, **tkwargs)
-                bounds[1] = 1
-                X = torch.rand(n_best, 21, **tkwargs)
-                # basic test
-                with mock.patch(
-                    "botorch.optim.initializers.draw_sobol_samples",
-                ) as mock_sobol:
-                    perturbed_X = sample_perturbed_subset_dims(
-                        X=X,
-                        n_discrete_points=n_discrete_points,
-                        qmc=False,
-                        sigma=1e-3,
-                        bounds=bounds,
-                    )
-                    mock_sobol.assert_not_called()
-                self.assertEqual(perturbed_X.shape, torch.Size([n_discrete_points, 21]))
-                self.assertTrue((perturbed_X >= 0).all())
-                self.assertTrue((perturbed_X <= 1).all())
-                # test qmc
-                with mock.patch(
-                    "botorch.optim.initializers.draw_sobol_samples",
-                    wraps=draw_sobol_samples,
-                ) as mock_sobol:
-                    perturbed_X = sample_perturbed_subset_dims(
-                        X=X,
-                        n_discrete_points=n_discrete_points,
-                        sigma=1e-3,
-                        bounds=bounds,
-                    )
-                    mock_sobol.assert_called_once()
-                self.assertEqual(perturbed_X.shape, torch.Size([n_discrete_points, 21]))
-                self.assertTrue((perturbed_X >= 0).all())
-                self.assertTrue((perturbed_X <= 1).all())
-                # for each point in perturbed_X compute the number of
-                # dimensions it has in common with each point in X
-                # and take the maximum number
-                max_equal_dims = (
-                    (perturbed_X.unsqueeze(0) == X.unsqueeze(1))
-                    .sum(dim=-1)
-                    .max(dim=0)
-                    .values
-                )
-                # check that at least one dimension is perturbed
-                self.assertTrue((20 - max_equal_dims >= 1).all())
-
     def test_sample_points_around_best(self):
         tkwargs = {"device": self.device}
         _bounds = torch.ones(2, 2)
